rename: touch_scale_widget.dart로 이름 변경
